requirements.txt
playwright
pandas
pytesseract
Pillow
requests